Refactor TOSA backend preprocess to use visitor pattern
