+++ backends/arm/operators/__init__.py
@@ -1,0 +1,22 @@
+# Copyright 2023 Arm Limited and/or its affiliates.
+#
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree.
+
+from . import (  # noqa
+    node_visitor,
+    op_add,
+    op_addmm,
+    op_avg_pool2d,
+    op_batch_norm,
+    op_clone,
+    op_conv2d,
+    op_dequant,
+    op_div,
+    op_get_item,
+    op_hardtanh,
+    op_permute,
+    op_quant,
+    op_softmax,
+    op_view,
+)
